создание отдельного модуля для вывода комментариев
