+++ js/rendering-full-photo.js
@@ -1,11 +1,10 @@
 import { isEscapeKey } from './util.js';
+import { displayComments } from './output-comments.js';
 
 const renderBigPhoto = (arrayPhotos) => {
   const bigPictureElement = document.querySelector('.big-picture');
   const bigPictureCloseElement = bigPictureElement.querySelector('.big-picture__cancel');
   const commentCountElement = document.querySelector('.social__comment-count');
-  const commentListElement = document.querySelector('.social__comments');
-  const commentElement = document.querySelector('.social__comment');
   const commentLoaderElement = document.querySelector('.comments-loader');
   const containerPhotoElement = document.querySelector('.pictures');
 
@@ -32,28 +31,6 @@ const renderBigPhoto = (arrayPhotos) => {
     bigPictureElement.classList.remove('hidden');
     document.body.classList.add('modal-open');
     document.addEventListener('keydown', onDocumentEscKeydown);
-  };
-
-  //создание одного комментария для списка
-  const getComment = ({ avatar, name, message }) => {
-    const comment = commentElement.cloneNode(true);
-    const socialPictureElement = comment.querySelector('.social__picture');
-
-    socialPictureElement.src = avatar;
-    socialPictureElement.alt = name;
-    comment.querySelector('.social__text').textContent = message;
-    return comment;
-  };
-
-  //создание списка комментариев
-  const displayComments = (comments) => {
-    const commentsFragmentElement = document.createDocumentFragment();
-    comments.forEach((element) => {
-      const newComment = getComment(element);
-      commentsFragmentElement.append(newComment);
-    });
-    commentListElement.innerHTML = '';
-    commentListElement.append(commentsFragmentElement);
   };
 
   //получаем данные для большого фото
